+++ plugin-dev/Source/Sentry/Private/Apple/SentrySubsystemApple.cpp
@@ -22,6 +22,8 @@
 
 #include "GenericPlatform/GenericPlatformOutputDevices.h"
 #include "HAL/FileManager.h"
+#include "Misc/App.h"
+#include "Misc/ConfigCacheIni.h"
 
 void SentrySubsystemApple::InitWithSettings(const USentrySettings* settings)
 {
@@ -29,8 +31,12 @@ void SentrySubsystemApple::InitWithSettings(const USentrySettings* settings)
 
 	[SENTRY_APPLE_CLASS(SentrySDK) startWithConfigureOptions:^(SentryOptions *options) {
 		options.dsn = settings->DsnUrl.GetNSString();
-		options.releaseName = settings->Release.GetNSString();
 		options.environment = settings->Environment.GetNSString();
+		options.enableAutoSessionTracking = settings->EnableAutoSessionTracking;
+		options.sessionTrackingIntervalMillis = settings->SessionTimeout;
+		options.releaseName = settings->OverrideReleaseName
+			? settings->Release.GetNSString()
+			: GetFormattedReleaseName().GetNSString();
 	}];
 
 	[SENTRY_APPLE_CLASS(SentrySDK) configureScope:^(SentryScope* scope) {
@@ -151,4 +157,36 @@ void SentrySubsystemApple::SetLevel(ESentryLevel level)
 	[SENTRY_APPLE_CLASS(SentrySDK) configureScope:^(SentryScope* scope) {
 		[scope setLevel:SentryConvertorsApple::SentryLevelToNative(level)];
 	}];
+}
+
+void SentrySubsystemApple::StartSession()
+{
+	[SENTRY_APPLE_CLASS(SentrySDK) startSession];
+}
+
+void SentrySubsystemApple::EndSession()
+{
+	[SENTRY_APPLE_CLASS(SentrySDK) endSession];
+}
+
+FString SentrySubsystemApple::GetFormattedReleaseName()
+{
+	FString FormattedReleaseName;
+
+#if PLATFORM_MAC
+	FString Version;
+	GConfig->GetString(TEXT("/Script/EngineSettings.GeneralProjectSettings"), TEXT("ProjectVersion"), Version, GGameIni);
+	if(!Version.IsEmpty())
+	{
+		FormattedReleaseName = FString::Printf(TEXT("%s@%s"), FApp::GetProjectName(), *Version);
+	}
+#elif PLATFORM_IOS
+	NSDictionary* infoDictionary = [[NSBundle mainBundle] infoDictionary];
+	FormattedReleaseName = FString([NSString stringWithFormat:@"%@@%@",
+		infoDictionary[@"CFBundleIdentifier"],
+		infoDictionary[@"CFBundleShortVersionString"]
+	]);
+#endif
+
+	return FormattedReleaseName;
 }